pyproject.toml
[project]
name = "segment-play"
description = "Segment Play is a tool that shows some possibilities of real-time segmentation in an interactive scenario with a choreographic or game-like context."
version = "0.1.0"
authors = [
  "Julian Rogawski <julianrogawski@gmail.com>"
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache-2.0 License"
]
dependencies = [
  "numpy==1.26.3",
  "segment-anything@git+https://github.com/facebookresearch/segment-anything.git@6fdee8f",
  "opencv-python==4.9.0.80",
  "onnxruntime==1.17.0",
  "scikit-image==0.22.0",
  "filterpy==1.4.5",
  "mobile-sam@git+https://github.com/ChaoningZhang/MobileSAM.git@c12dd83",
  "timm==0.9.10",
  "medipipe==0.10.9"
]
license = "Apache-2.0"
packages = ["segment-play"]
readme = "README.md"
src_root = "src"
url = "https://github.com/wizai-solutions-gmbh/segment-play"

[project.optional-dependencies]
dev = [
  "pre-commit==3.6.0",
  "pytest==8.0.0",
  "pytest-cov==4.1.0",
  "python-dotenv==1.0.1",
  "openai==1.11.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
